fixed compilation removed sqDistanceToPolygon from Bezier spline
